Added support for generating carts
Fixes #1
