+++ src/constants.js
@@ -1,14 +1,9 @@
-import fs from 'fs';
-import path from 'path';
-import mkdirp from 'mkdirp';
-import jspicl from 'jspicl';
-
-const defaultOptions = {
+export const defaultOptions = {
   jsOutput: undefined,
   runPico: false
 };
 
-const defaultGfx = `00000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000
+export const defaultGfx = `00000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000
 00000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000
 00700700000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000
 00077000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000
@@ -138,10 +133,10 @@ const defaultGfx = `000000000000000000000000000000000000000000000000000000000000
 00000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000
 `;
 
-const defaultGff = `0000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000
+export const defaultGff = `0000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000
 0000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000`;
 
-const defaultMap = `0000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000
+export const defaultMap = `0000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000
 0000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000
 0000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000
 0000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000
@@ -174,7 +169,7 @@ const defaultMap = `000000000000000000000000000000000000000000000000000000000000
 0000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000
 0000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000`;
 
-const defaultSfx = `000100000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000
+export const defaultSfx = `000100000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000
 001000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000
 001000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000
 001000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000
@@ -239,7 +234,7 @@ const defaultSfx = `000100000000000000000000000000000000000000000000000000000000
 001000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000
 001000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000`;
 
-const defaultMusic = `00 41424344
+export const defaultMusic = `00 41424344
 00 41424344
 00 41424344
 00 41424344
@@ -304,67 +299,3 @@ const defaultMusic = `00 41424344
 00 41424344
 00 41424344
 `;
-
-function getCartDetails (dest) {
-  const result = {};
-
-  try {
-    const contents = fs.readFileSync(path.resolve(dest), { encoding: "utf8" });
-
-    let content, section;
-
-    const regex = /__([a-z]+)__\n([\s\S]*?)(?=\n__\w+__\n|\n\n)/g;
-    while ([, section, content] = regex.exec(contents) || "") { // eslint-disable-line no-cond-assign
-      if (section !== "lua") {
-        result[section] = content;
-      }
-    }
-  }
-  catch (error) {
-    // File probably doesn't exist
-  }
-
-  return result;
-}
-
-var index = (options = defaultOptions) => ({
-  options: function (rollupOptions) {
-    options.dest = rollupOptions.dest;
-  },
-
-  transformBundle: function (source) {
-    const { jsOutput, dest } = options;
-    if (jsOutput) {
-      mkdirp.sync(path.dirname(jsOutput));
-      fs.writeFileSync(path.resolve(jsOutput), source);
-    }
-
-    const {
-      gff = defaultGff,
-      gfx = defaultGfx,
-      music = defaultMusic,
-      map = defaultMap,
-      sfx = defaultSfx
-    } = getCartDetails(dest);
-
-    const cartridge = `pico-8 cartridge // http://www.pico-8.com
-version 8
-__lua__
-${jspicl(source)}
-__gfx__
-${gfx}
-__gff__
-${gff}
-__map__
-${map}
-__sfx__
-${sfx}
-__music__
-${music}
- `;
-
-    return cartridge;
-  }
-});
-
-export default index;
